+++ src/renderer/components/swap/Swap.tsx
@@ -875,7 +875,7 @@ export const Swap = ({
           const amount = new CryptoAmount(convertBaseAmountDecimal(amountToSwapMax1e8, sourceAssetDecimal), sourceAsset)
           const address = destinationAddress
           const fromAdd = sourceWalletAddress
-          const toleranceBps = isStreaming ? 10000 : undefined // convert to basis points
+          const toleranceBps = isStreaming ? 10000 : slipTolerance * 100 // convert to basis points
           return {
             fromAsset: fromAsset,
             destinationAsset: destinationAsset,
@@ -896,6 +896,7 @@ export const Swap = ({
       amountToSwapMax1e8,
       sourceAssetDecimal,
       isStreaming,
+      slipTolerance,
       applyBps
     ]
   )
@@ -946,7 +947,7 @@ export const Swap = ({
             destinationAsset: targetAsset,
             fromAddress: sourceWalletAddress,
             amount: new CryptoAmount(convertBaseAmountDecimal(amountToSwapMax1e8, sourceAssetDecimal), sourceAsset),
-            toleranceBps: undefined,
+            toleranceBps: slipTolerance * 100,
             affiliateAddress: ASGARDEX_THORNAME,
             affiliateBps: applyBps
           }
@@ -1267,6 +1268,18 @@ export const Swap = ({
       })
     )
   }, [oQuote, disableSlippage, swapResultAmountMax, zeroTargetBaseAmountMax1e8])
+
+  const swapLimitMaya1e8: O.Option<BaseAmount> = useMemo(() => {
+    return FP.pipe(
+      oQuoteMaya,
+      O.chain((txDetails) => {
+        // Disable slippage protection temporary for Ledger/BTC (see https://github.com/thorchain/asgardex-electron/issues/2068)
+        return !disableSlippage && swapResultAmountMax.baseAmount.gt(zeroTargetBaseAmountMax1e8)
+          ? O.some(Utils.getSwapLimit1e8(txDetails.memo))
+          : O.none
+      })
+    )
+  }, [oQuoteMaya, disableSlippage, swapResultAmountMax, zeroTargetBaseAmountMax1e8])
 
   const oSwapParams: O.Option<SwapTxParams> = useMemo(
     () => {
@@ -1718,7 +1731,11 @@ export const Swap = ({
       setIsStreaming(streamingIntervalValue !== 0)
     }
     const tipFormatter =
-      slider === 0 ? 'Caution tx could be refunded' : `${streamingIntervalValue} Block interval between swaps`
+      dex === 'THOR'
+        ? slider === 0
+          ? 'Caution tx could be refunded'
+          : `${streamingIntervalValue} Block interval between swaps`
+        : `Mayachain does not support streaming yet`
     const labelMin = slider <= 0 ? `Limit Swap` : `` || slider < 50 ? 'Time Optimised' : `Price Optimised`
 
     return (
@@ -1733,10 +1750,12 @@ export const Swap = ({
           tipFormatter={() => `${tipFormatter} `}
           labels={[`${labelMin}`, `${streamingInterval}`]}
           tooltipPlacement={'top'}
+          error={dex === 'MAYA'}
+          disabled={dex === 'MAYA'}
         />
       </div>
     )
-  }, [slider, streamingInterval])
+  }, [dex, slider, streamingInterval])
 
   // Streaming Quantity slider
   const renderStreamerQuantity = useMemo(() => {
@@ -1748,7 +1767,7 @@ export const Swap = ({
     let toolTip: string
     if (streamingInterval === 0) {
       quantityLabel = [`Limit swap`]
-      toolTip = `No Streaming interval set`
+      toolTip = dex === 'THOR' ? `No Streaming interval set` : `Mayachain does not support streaming yet`
     } else {
       quantityLabel = quantity === 0 ? [`Auto swap count`] : [`Sub swaps`, `${quantity}`]
       toolTip =
@@ -1770,10 +1789,12 @@ export const Swap = ({
           included={false}
           labels={quantityLabel}
           tooltipPlacement={'top'}
+          error={dex === 'MAYA'}
+          disabled={dex === 'MAYA'}
         />
       </div>
     )
-  }, [maxStreamingQuantity, streamingQuantity, streamingInterval])
+  }, [streamingQuantity, streamingInterval, maxStreamingQuantity, dex])
 
   // swap expiry progress bar
   useEffect(() => {
@@ -1835,6 +1856,7 @@ export const Swap = ({
         withLabel={true}
         labels={[`${streamerComparison}`, ``]}
         tooltipPlacement={'top'}
+        hasError={!isStreaming}
       />
     )
   }, [isStreaming, swapSlippage, swapStreamingSlippage])
@@ -2188,13 +2210,22 @@ export const Swap = ({
   // Label: Min amount to swap (<= 1e8)
   const swapMinResultLabel = useMemo(() => {
     // for label we do need to convert decimal back to original decimal
-    const amount: BaseAmount = FP.pipe(
-      swapLimit1e8,
-      O.fold(
-        () => baseAmount(0, targetAssetDecimal) /* assetAmount1e8 */,
-        (limit1e8) => convertBaseAmountDecimal(limit1e8, targetAssetDecimal)
-      )
-    )
+    const amount: BaseAmount =
+      dex === 'THOR'
+        ? FP.pipe(
+            swapLimit1e8,
+            O.fold(
+              () => baseAmount(0, targetAssetDecimal) /* assetAmount1e8 */,
+              (limit1e8) => convertBaseAmountDecimal(limit1e8, targetAssetDecimal)
+            )
+          )
+        : FP.pipe(
+            swapLimitMaya1e8,
+            O.fold(
+              () => baseAmount(0, targetAssetDecimal) /* assetAmount1e8 */,
+              (limit1e8) => convertBaseAmountDecimal(limit1e8, targetAssetDecimal)
+            )
+          )
 
     const amountMax1e8 = max1e8BaseAmount(amount)
 
@@ -2205,7 +2236,7 @@ export const Swap = ({
           amount: baseToAsset(amountMax1e8),
           trimZeros: true
         })}`
-  }, [disableSlippage, swapLimit1e8, targetAssetDecimal, targetAsset])
+  }, [dex, swapLimit1e8, swapLimitMaya1e8, disableSlippage, targetAsset, targetAssetDecimal])
 
   const uiApproveFeesRD: UIFeesRD = useMemo(
     () =>
@@ -2884,7 +2915,9 @@ export const Swap = ({
                       <div>{intl.formatMessage({ id: 'swap.slip.title' })}</div>
                       <div>
                         {formatAssetAmountCurrency({
-                          amount: priceAmountToSwapMax1e8.assetAmount.times(swapSlippage / 100), // Find the value of swap slippage
+                          amount: priceAmountToSwapMax1e8.assetAmount.times(
+                            (swapSlippage > 0 ? swapSlippage : slipTolerance) / 100
+                          ), // Find the value of swap slippage
                           asset: priceAmountToSwapMax1e8.asset,
                           decimal: isUSDAsset(priceAmountToSwapMax1e8.asset) ? 2 : 6,
                           trimZeros: !isUSDAsset(priceAmountToSwapMax1e8.asset)